### 2023-07-01 #5610 @editorialbot
[REVIEW]: cellanneal: A user-friendly deconvolution software for omics data
**Submitting author:** @libuchauer (<a href="http://orcid.org/0000-0002-4722-8390">Lisa Buchauer</a>)
**Repository:** https://github.com/LiBuchauer/cellanneal
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
