+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={1--14},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41467-020-19015-1}
 }
 
 @article{sturm2019benchmarking,
@@ -43,7 +44,8 @@
   number={1},
   pages={1--9},
   year={2019},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41467-018-08023-x}
 }
 
 @article{erdmann2021likelihood,
@@ -54,7 +56,8 @@
   number={10},
   pages={1794--1806},
   year={2021},
-  publisher={Cold Spring Harbor Lab}
+  publisher={Cold Spring Harbor Lab},
+  doi={10.1101/gr.272344.120}
 }
 
 @article{jew2020accurate,
@@ -76,7 +79,8 @@
   number={5},
   pages={e1006976},
   year={2019},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pcbi.1006976}
 }
 
 @article{newman2015robust,
@@ -87,7 +91,8 @@
   number={5},
   pages={453--457},
   year={2015},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nmeth.3337}
 }
 
 @article{newman2019determining,
@@ -99,15 +104,24 @@
   pages={773--782},
   year={2019},
   publisher={Nature Publishing Group}
+  doi={10.1038/s41587-019-0114-2}
 }
 
 @article{aliee2021autogenes,
-  title={AutoGeneS: Automatic gene selection using multi-objective optimization for RNA-seq deconvolution},
-  author={Aliee, Hananeh and Theis, Fabian J},
-  journal={Cell Systems},
-  year={2021},
-  publisher={Elsevier}
-}
+    title = {AutoGeneS: Automatic gene selection using multi-objective optimization for RNA-seq deconvolution},
+    journal = {Cell Systems},
+    volume = {12},
+    number = {7},
+    pages = {706-715.e4},
+    year = {2021},
+    issn = {2405-4712},
+    doi = {10.1016/j.cels.2021.05.006},
+    url = {https://www.sciencedirect.com/science/article/pii/S2405471221001927},
+    author = {Aliee, Hananeh and Theis, Fabian J.},
+    keywords = {bulk deconvolution, bulk RNA-seq, single-cell RNA-seq, multi-objective optimization, feature selection, marker genes},
+    abstract = {Summary
+    Knowing cell-type proportions in a tissue is very important to identify which cells or cell types are targeted by a disease or perturbation. Hence, several deconvolution methods have been proposed to infer cell-type proportions from bulk RNA samples. Their performance with noisy reference profiles and closely correlated cell types highly depends on the set of genes undergoing deconvolution. In this work, we introduce AutoGeneS, a platform that automatically extracts discriminative genes and reveals the cellular heterogeneity of bulk RNA samples. AutoGeneS requires no prior knowledge about marker genes and selects genes by simultaneously optimizing multiple criteria: minimizing the correlation and maximizing the distance between cell types. AutoGeneS can be applied to reference profiles from various sources like single-cell experiments or sorted cell populations. Ground truth cell proportions analyzed by flow cytometry confirmed the accuracy of AutoGeneS in identifying cell-type proportions. AutoGeneS is available for use via a standalone Python package (https://github.com/theislab/AutoGeneS).}
+    }
 
 @article{li2016comprehensive,
[…]
