### 2023-07-02 a06d614
added more missing DOIs
