+++ paper/paper.bib
@@ -33,7 +33,8 @@
   volume={6},
   pages={e26476},
   year={2017},
-  publisher={eLife Sciences Publications Limited}
+  publisher={eLife Sciences Publications Limited},
+  doi={10.7554/eLife.26476}
 }
 
 @article{wang2019bulk,
@@ -68,7 +69,8 @@
   number={1},
   pages={1--11},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41467-020-15816-6}
 }
 
 @article{hao2019fast,
@@ -103,7 +105,7 @@
   number={7},
   pages={773--782},
   year={2019},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
   doi={10.1038/s41587-019-0114-2}
 }
 
@@ -131,7 +133,8 @@
   number={1},
   pages={1--16},
   year={2016},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.1186/s13059-016-1028-7}
 }
 
 @article{schwanhausser2011global,
@@ -142,7 +145,8 @@
   number={7347},
   pages={337--342},
   year={2011},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={Global quantification of mammalian gene expression control}
 }
 
 @article{kirkpatrick1983optimization,
@@ -153,7 +157,8 @@
   number={4598},
   pages={671--680},
   year={1983},
-  publisher={American association for the advancement of science}
+  publisher={American association for the advancement of science},
+  doi={10.1126/science.220.4598.671}
 }
 
 @ARTICLE{2020SciPy-NMeth,
@@ -257,7 +262,8 @@
   number={5},
   pages={e3002124},
   year={2023},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pbio.3002124}
 }
 
 @article{berkova2022terminal,
