### 2023-08-19 5a6a689
Update README.md
JOSS review - typo
