### 2023-08-22 943b74e
make clear that both pip and conda installs require pip install .
