### 2023-08-22 3b8081e
Added environment.yml and requirements.txt files and corresponding explanations
