+++ requirements.txt
@@ -1,0 +1,7 @@
+numpy == 1.24.3
+scipy == 1.9.3
+matplotlib == 3.7.1
+pandas == 1.5.3
+seaborn == 0.12.2
+xlrd == 2.0.1
+openpyxl == 3.0.10
