### 2023-08-22 #5610 @LiBuchauer
Thanks @ManavalanG, excellent points. I have started working on it as documented below.

> ## Installation
> *

>- [x]  Installation requires several dependencies such as `numpy`, etc., but their versions are not pinned. I would highly recommend pinning them (eg. `numpy>=1.24` to eliminate version related errors. --> addressed in 3b8081efe48a6d69183f7a184f6aa52844da5a47
> * [x]  _Minor suggestion._ It would be a good idea to specify the dependencies in `requirements.txt` file for pip-based installation and using [`environment.yml` file for conda-based installation](https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-file-manually). These files would also make it easier to specify the dependency versions. --> addressed in 3b8081efe48a6d69183f7a184f6aa52844da5a47
>
> ## Documentation
> * [x]  _Minor suggestion._`pip install .` needs to be executed irrespective of whether dependencies were installed via conda or pip. However It was unclear to me this step was needed when I chose the conda route. A minor reorganization would help. --> addressed in 285b24c6792bcb8b98b287a039a901dd45239768
> * [ ]  A snippet in `README.md` - "The repository contains example data from a publication on liver cancer microenvironments at examples/example_data/". Please cite the publication here. It is cited in file `examples/cellanneal_quickstart.ipynb`, but mentioning it in the readme doc would greatly help the users.
> * [ ]  Fulfill this requirement from JOSS's checklist on documentation: "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". This is already accomplished by the included file `CONTRIBUTING.md` but mose users would likely not know about this file. So, linking it in the readme doc would make it easier for folks to come across the contributing and support guidelines.
>
> ## Functionality
> * [ ]  Regarding performance, manuscript states that "Its typical processing time for one mixture sample is below one minute on a desktop machine". However, in my testing the tool on Macbook Pro 2019 (2.4GHz 8-core processor, 32GB mem), it took ~2mins. Command I used: `time cellanneal examples/example_data/mixture_data_liver_tumor.csv examples/example_data/signature_data_human_liver.csv output_directory`. Time output on my Mac: `Time output: "674.31s user 179.48s system 738% cpu 1:55.67 total"`. I ran the command for a total of 3 times, and it took ~2mins every time. Please complete the following:
>
>   * [ ]  Mention the specs of the desktop machine where testing was performed
>   * [x]  Confirm the time taken to run the tool. --> I take the liberty to check this, because the text file contains 5 columns, i.e. 5 mixture samples, meaning that the statement with less than 1 minute per mixture sample holds on your machine
