### 2023-08-25 #5610 @ManavalanG
Below is the continuation of my review.

## Manuscript

* [ ] Title is really broad considering content of the manuscript. It discusses application in the field of transcriptomics, whereas title says `software for omics data`. Justification on how it applies to other type of omics data need to be included, or title needs to be modified to reflect the manuscript's content (ie. transciptomics data).
* [ ] The end of second paragraph in the `summary` section mentions two major optimization algorithms (least squares regression and support vector regression) used to solve the problem, and then the third paragraph discusses the challenges of using least squares regression. However such description on the topic of support vector regression is missing, and adding a note on them would guide the reader. Also, mentioning how the algorithm used in `cellanneal` (spearman’s rank correlation coefficient) fits in between the above two algorithms would be helpful.
* [ ] _Minor_. Mentioning dataset used to obtain the figures used in manuscript could be helpful.

## Tool installation

* [ ] Documentation is missing info on how to install a particular version of `cellanneal`. For example, how to install [`v1.0.0`](https://github.com/LiBuchauer/cellanneal/releases/tag/v1.0.0)?

**Note:** In my initial feedback provided on Aug 21, I had tagged my comment about tool dependency version and conda environment definition  as `minor suggestion` but upon further reflection, I removed that tag. Having tool versions defined would help with reliable installation and reproducibility. Please let me know if you have any questions :)
