+++ README.md
@@ -117,7 +117,7 @@ For more information on how to use the GUI, see [Using the graphical software](#
 
 #### 2c. Testing your installation
 
-The repository contains example data from a publication on liver cancer microenvironments at `examples/example_data/`. Use this data to test your installation as described below to produce a set of outputs, consisting of tabular data and figures, as shown in `examples/example_output/`. 
+The repository contains example data from a publication on liver cancer microenvironments ([Massalha et al., 2020](https://doi.org/10.15252/msb.20209682)) at `examples/example_data/`. Use this data to test your installation as described below to produce a set of outputs, consisting of tabular data and figures, as shown in `examples/example_output/`. 
 
 #### python package
 Run the quick start jupyter notebook at `examples/cellanneal_quickstart.ipynb` or copy the code from the [web rendering of the notebook](https://github.com/LiBuchauer/cellanneal/blob/master/examples/cellanneal_quickstart.ipynb) into your python console and execute it there.
