### 2023-08-28 a944bc8
Link to CONTRIBUTING.md from README.md
