### 2023-08-28 9275cca
Included specs of the testing machine in the corresponding paper section.
