+++ paper/paper.md
@@ -40,7 +40,7 @@ Here, we introduce `cellanneal`, a python-based software for deconvolving bulk R
 
 Making sense of bulk RNA sequencing datasets often requires analysis of the cell type composition of the samples. This is particularly relevant in clinical samples that analyze the transcriptome of tissues or tumors which consist of epithelial, stromal and immune cell types. In parallel, publicly available single-cell data sets enable precise characterization of the expression signature of multiple individual cell types. However, software tools for computational bulk deconvolution are often slow, non-robust and not easy to use. Some existing methods address the aspect of user-friendliness by providing graphical web interfaces, but submitting sensitive medical data to an external web server is not always compatible with privacy legislation.
 
-To address these challenges, we have developed `cellanneal`, a deconvolution approach that uses Spearman’s rank correlation coefficient between synthetic and bulk gene expression vectors as the optimization procedure’s objective function. Because this correlation measure is calculated from ranks rather than absolute data values, each gene influences the optimization result to a similar extent. Users are encouraged to include as many informative genes as possible in the analysis. `cellanneal` optimizes cell type fractions by simulated annealing, a flexible, rapid and robust algorithm for global optimization [@kirkpatrick1983optimization; @2020SciPy-NMeth]. `cellanneal` can be used as a python package, via its command line interface or via a user-friendly graphical software which runs locally. Its typical processing time for one mixture sample is below one minute on a desktop machine.
+To address these challenges, we have developed `cellanneal`, a deconvolution approach that uses Spearman’s rank correlation coefficient between synthetic and bulk gene expression vectors as the optimization procedure’s objective function. Because this correlation measure is calculated from ranks rather than absolute data values, each gene influences the optimization result to a similar extent. Users are encouraged to include as many informative genes as possible in the analysis. `cellanneal` optimizes cell type fractions by simulated annealing, a flexible, rapid and robust algorithm for global optimization [@kirkpatrick1983optimization; @2020SciPy-NMeth]. `cellanneal` can be used as a python package, via its command line interface or via a user-friendly graphical software which runs locally. Its typical processing time for one mixture sample is below one minute on a desktop machine (MacBook Pro 2020, 2.3 GHz Quad-Core Intel Core i7, 16 GB RAM).
 
 # Availability and Features
 
