+++ paper/paper.md
@@ -32,9 +32,9 @@ Single-cell sequencing methods enable precise characterization of gene expressio
 
 For these reasons, computational deconvolution methods are gaining popularity in basic and clinical research. Computational deconvolution approaches infer the cell type proportions constituting a given bulk RNA sample based on separately obtained cell type reference data. Several computational deconvolution methods have been developed in the last decade and have contributed to our understanding of tissue composition [@cobos2020benchmarking; @sturm2019benchmarking]. Generally, during deconvolution, the computational mixture is constructed from a set of cell type fractions and reference gene expression vectors for each of the participating cell types, most commonly derived from single-cell data. The cell type fractions are then iteratively changed until agreement between the _in silico_ gene expression vector and the observed bulk sample gene expression vector is optimal by a measure of choice. Here, published methods rely almost exclusively on minimizing the sum of squared residuals between bulk and computationally mixed vectors. Algorithms for such optimization problems are readily available and include variants of least squares regression (e.g. weighted least squares regression [@racle2017simultaneous], non-negative least squares regression [@wang2019bulk; @jew2020accurate] or least trimmed squares [@hao2019fast]) and support vector regression [@newman2015robust; @newman2019determining]. 
 
-However, least squares-based optimization is faced with a particular challenge in bulk RNAseq deconvolution because of the highly skewed nature of mRNA copy number distributions, ranging from less than 1 to more than 10,000 average mRNA copies per cell [@li2016comprehensive; @schwanhausser2011global]. In such settings, optimization results may be strongly influenced by few highly expressed genes and are thus not robust to noise or platform effects influencing the readout of these genes. Identifying the right genes for deconvolution becomes a task in itself [@aliee2021autogenes]. As a result, deconvolution methods may yield inferred mixed gene expression vectors that do not correlate well with measured bulk gene expression. 
+However, least squares-based optimization is faced with a particular challenge in bulk RNAseq deconvolution because of the highly skewed nature of mRNA copy number distributions, ranging from less than 1 to more than 10,000 average mRNA copies per cell [@li2016comprehensive; @schwanhausser2011global]. In such settings, optimization results may be strongly influenced by few highly expressed genes and are thus not robust to noise or platform effects influencing the readout of these genes. Support vector regression based models like CIBERSORT [@newman2015robust] perform gene feature selection out of a user-defined signature gene list, the contents of which can strongly affect the cell proportion estimates. Overall, identifying the right genes for deconvolution becomes a task in itself [@aliee2021autogenes]. As a result, deconvolution methods may yield inferred mixed gene expression vectors that do not correlate well with measured bulk gene expression.
 
-Here, we introduce `cellanneal`, a python-based software for deconvolving bulk RNA sequencing data. `cellanneal` relies on the optimization of Spearman's rank correlation coefficient between experimental and computational mixture gene expression vectors using simulated annealing. `cellanneal` can be used as a python package or via a command line interface, but importantly also provides a simple graphical user interface which is distributed as a single executable file for user convenience.
[…]
