+++ paper/paper.bib
@@ -275,3 +275,40 @@
   publisher={Cold Spring Harbor Laboratory},
   doi={10.1101/2022.11.11.516138}
 }
+
+
+@ARTICLE{massalha2020liver,
+  title    = "A single cell atlas of the human liver tumor microenvironment",
+  author   = "Massalha, Hassan and Bahar Halpern, Keren and Abu-Gazala, Samir
+              and Jana, Tamar and Massasa, Efi E and Moor, Andreas E and
+              Buchauer, Lisa and Rozenberg, Milena and Pikarsky, Eli and Amit,
+              Ido and Zamir, Gideon and Itzkovitz, Shalev",
+  abstract = "Malignant cell growth is fueled by interactions between tumor
+              cells and the stromal cells composing the tumor microenvironment.
+              The human liver is a major site of tumors and metastases, but
+              molecular identities and intercellular interactions of different
+              cell types have not been resolved in these pathologies. Here, we
+              apply single cell RNA-sequencing and spatial analysis of
+              malignant and adjacent non-malignant liver tissues from five
+              patients with cholangiocarcinoma or liver metastases. We find
+              that stromal cells exhibit recurring, patient-independent
+              expression programs, and reconstruct a ligand-receptor map that
+              highlights recurring tumor-stroma interactions. By combining
+              transcriptomics of laser-capture microdissected regions, we
+              reconstruct a zonation atlas of hepatocytes in the non-malignant
+              sites and characterize the spatial distribution of each cell type
+              across the tumor microenvironment. Our analysis provides a
+              resource for understanding human liver malignancies and may
+              expose potential points of interventions.",
+  journal  = "Mol. Syst. Biol.",
+  volume   =  16,
+  number   =  12,
+  pages    = "e9682",
+  month    =  dec,
+  year     =  2020,
+  keywords = "human cell atlas; liver cancer; single cell RNAseq; spatial
+              transcriptomics; tumor-stroma interactions",
+  language = "en"
+}
+
+
