### 2023-10-24 #5610 @LiBuchauer
Hi @ManavalanG , I addressed the remaining 3 points about the manuscript as detailed below, thank you for your input. Please let me know if anything else is lacking.

> Here are my further comments, following [my earlier feedback/comments](https://github.com/openjournals/joss-reviews/issues/5610#issuecomment-1687374013).
>
> ## Manuscript
> * [ X]  Title is really broad considering content of the manuscript. It discusses application in the field of transcriptomics, whereas title says `software for omics data`. Justification on how it applies to other type of omics data need to be included, or title needs to be modified to reflect the manuscript's content (ie. transciptomics data). --> I changed it as requested.
> * [X]  The end of second paragraph in the `summary` section mentions two major optimization algorithms (least squares regression and support vector regression) used to solve the problem, and then the third paragraph discusses the challenges of using least squares regression. However such description on the topic of support vector regression is missing, and adding a note on them would guide the reader. Also, mentioning how the algorithm used in `cellanneal` (spearman’s rank correlation coefficient) fits in between the above two algorithms would be helpful. --> I added one sentence about SVR method problems, and also outlined what cellanneal does differently.
> * [X]  _Minor_. Mentioning dataset used to obtain the figures used in manuscript could be helpful. --> Done
>
> ## Tool installation
> * [ ]  Documentation is missing info on how to install a particular version of `cellanneal`. For example, how to install [`v1.0.0`](https://github.com/LiBuchauer/cellanneal/releases/tag/v1.0.0)? _[Update Oct 2] This is minor and recommended but does not need to be included._ --> I choose not to do this in the interest of time and also because there are currently no several versions :) hope it's okay
>
> **Note:** In my initial feedback provided on Aug 21, I had tagged my comment about tool dependency version and conda environment definition as `minor suggestion` but upon further reflection, I removed that tag. Having tool versions defined would help with reliable installation and reproducibility. Please let me know if you have any questions :) --> I had already adressed this
