### 2023-11-04 #5610 @ritika-giri
Sorry for the delay @jmschrei  - I have 2 major comments:
1. Authors do not mention other publications, or packages in R / Python that implement simulated annealing algorithms for cell mixture deconvolution. These should be clearly referenced in order to establish the broader context of the field and how their work fits into it.
2. I suspect that this package does not fulfill the criteria of substantial scholarly effort, although I am more than happy to be corrected by the authors or editor.  As I understand, the file dual_annealing.py and general.py contain the algorithms. Of these, dual_annealing.py is from scipy's codebase with no modifications.  The file general.py has two major functions - to find highly variable genes and deconvolution functions. Of these, the code for finding variable genes is from scanpy's implementation of the same, while deconvolve functions are calling the dual annealing code (originally in scipy). I would like the authors to clearly state the algorithmic modifications that are novel in their package, or their justifications for substantial scholarly effort.
