### 2023-11-06 #5610 @LiBuchauer
Hi good morning @ritika-giri,

regarding the first point, to my best knowledge there are no other packages that use simulated annealing as an optimization procedure and I consider this an original idea we had. Obviously I will cite and discuss any such paper that you can name. The only one I could find just now which mentions both the term "cell type deconvolution" and "simulated annealing” is this one https://doi.org/10.1093/nargab/lqab056 but it does not use simulated annealing for cell type deconvolution. A lot of the published methods rely on least squares like distance metrics because of availability of fast optimization algorithms for them. In many omics contexts however, non-parametric methods have proven to be more stable, and thus, in cellanneal, we implemented a solution with Spearman’s rank correlation coefficient as a distance function and simulated annealing as an optimization procedure.

This also brings us to the second point. dual_annealing.py, as you note, is from scipy, which is clearly cited and marked. The reason that I copied the code over is that an important part of cellanneal is the GUI which comes as a single executable, and in order to keep this slim and reduce its start-up time, I wanted to remove scipy from the dependancies. Around this simulated annealing implementation, we provide a pipeline for importing bulk and signature data, extracting highly variable genes (yes, via the scanpy implementation, as is clearly marked), running the optimization and returning relevant plots as well as tabular results. The functionality is accessible via python, cli or GUI. We find our method to perform well and fast.

Overall, in my view, the most important points are 1) we present a new idea for performing cell type deconvolution based on a non-parametric distance function and 2) (maybe more relevant for JOSS), we provide this as an easy-to-use software which can also be employed by non-coding scientists and runs locally. So far, cellanneal has been cited by 3 peer reviewed publications (https://scholar.google.com/scholar?cites=10240263254458724390&as_sdt=2005&sciodt=0,5&hl=en), including two wholly unrealted to the cellanneal authors, and I know from people seeking support via email that is has also found its way into some biotech companies. It’s probably not worth anything, but the work that went into this is certainly (much) more than 3 months full time, though of course not all was spent on the code, but on conceiving and testing the method together with experimental biologists.

@ritika-giri hope this clarifies it a bit & @jmschrei hope you can make a call based on this explanation.
