### 2023-11-08 #5610 @jmschrei
In general, I think it's better to import functions from other packages (even big ones like scipy) so that (1) they get the credit they deserve for writing the original algorithm and (2) any upstream improvements in performance can make their way into your package without any effort on your part. That being said, it's not a strict requirement and the authors do see to clearly state where the code is from.

I personally believe that the substantial scholarly effort criteria have been met. Remember that substantial effort at JOSS focuses more on the development of the code rather than algorithmic novelty -- though it does need to fill a niche. @ritika-giri if you know of any specific publications using simulated annealing you think they should mention, I agree that they should include them, even if they aren't explicitly optimizing Spearman R.
