### 2023-12-12 #5610 @jmschrei
@LiBuchauer can you provide the DOI for the paper (e.g., from Zenodo) and the version of the code associated with this submission? And can you check out that invalid DOI?
