### 2023-12-19 e4dd5e5
Update paper.bib
fixed doi for schwanhaeusser
