+++ paper/paper.bib
@@ -146,7 +146,7 @@
   pages={337--342},
   year={2011},
   publisher={Nature Publishing Group},
-  doi={Global quantification of mammalian gene expression control}
+  doi={10.1038/nature10098}
 }
 
 @article{kirkpatrick1983optimization,
