### 2023-12-19 #5610 @LiBuchauer
@jmschrei the code release including the latest changes is [v1.1.0](https://github.com/LiBuchauer/cellanneal/releases/tag/v1.1.0). The zenodo DOI associated with the release is [10.5281/zenodo.10405043](https://doi.org/10.5281/zenodo.10405043). Hope I did it correctly, if not, let me know!
