### 2023-12-19 #5610 @jmschrei
I noticed a few issues in the references. Would you mind fixing:

(1) Capitalization on L124 in "PLoS"
(2) Capitalization on L49 in "www"
(3) Capitalization on L163 in "eLife"
(4) Authors on L172 as "Pandas Development Team"

Thanks!
