+++ paper/paper.bib
@@ -188,7 +188,7 @@
  @article{lundh1999introduction,
  title={An introduction to tkinter},
  author={Lundh, Fredrik},
- journal={URL: www.pythonware.com/library/tkinter/introduction/index.htm},
+ journal={https://www.pythonware.com/library/tkinter/introduction/index.htm},
  year={1999}
 }
 
