### 2023-12-20 6793703
Update paper.bib
remove tkinter citation as the link is broken
