+++ paper/paper.bib
@@ -185,12 +185,6 @@
   doi     = {10.1038/s41592-019-0686-2},
 }
 
- @article{lundh1999introduction,
- title={An introduction to tkinter},
- author={Lundh, Fredrik},
- journal={https://www.pythonware.com/library/tkinter/introduction/index.htm},
- year={1999}
-}
 
 @Article{         harris2020array,
  title         = {Array programming with {NumPy}},
