### 2023-12-20 28072fa
Update paper.bib
pandas team as author spelled out in full
