+++ paper/paper.bib
@@ -210,7 +210,7 @@
 }
 
 @software{reback2020pandas,
-    author       = {Pandas Development Team},
+    author       = {{Pandas Development Team}},
     title        = {pandas-dev/pandas: Pandas},
     month        = feb,
     year         = 2020,
