+++ paper/paper.bib
@@ -272,12 +272,12 @@
 
 
 @ARTICLE{massalha2020liver,
-  title    = "A single cell atlas of the human liver tumor microenvironment",
-  author   = "Massalha, Hassan and Bahar Halpern, Keren and Abu-Gazala, Samir
+  title    = {A single cell atlas of the human liver tumor microenvironment},
+  author   = {Massalha, Hassan and Bahar Halpern, Keren and Abu-Gazala, Samir
               and Jana, Tamar and Massasa, Efi E and Moor, Andreas E and
               Buchauer, Lisa and Rozenberg, Milena and Pikarsky, Eli and Amit,
-              Ido and Zamir, Gideon and Itzkovitz, Shalev",
-  abstract = "Malignant cell growth is fueled by interactions between tumor
+              Ido and Zamir, Gideon and Itzkovitz, Shalev},
+  abstract = {Malignant cell growth is fueled by interactions between tumor
               cells and the stromal cells composing the tumor microenvironment.
               The human liver is a major site of tumors and metastases, but
               molecular identities and intercellular interactions of different
@@ -293,16 +293,14 @@
               sites and characterize the spatial distribution of each cell type
               across the tumor microenvironment. Our analysis provides a
               resource for understanding human liver malignancies and may
-              expose potential points of interventions.",
-  journal  = "Mol. Syst. Biol.",
-  volume   =  16,
-  number   =  12,
-  pages    = "e9682",
-  month    =  dec,
-  year     =  2020,
-  keywords = "human cell atlas; liver cancer; single cell RNAseq; spatial
-              transcriptomics; tumor-stroma interactions",
-  language = "en"
+              expose potential points of interventions.},
+  journal  = {Mol. Syst. Biol.},
+  volume   =  {16},
+  number   =  {12},
+  pages    = {e9682},
+  month    =  {dec},
+  year     =  {2020},
+  doi={10.15252/msb.20209682}
 }
 
 
