### 2024-01-12 #5610 @LiBuchauer
Hi @Kevin-Mattheus-Moerman , thanks, I added the doi and changed the license type on zenodo.
